chore: exception handler moved to core
